1.생성 부분 분할 2.uint32_t -> uint16_t로 바꾸어 렌더링 성공 / TODO : 원인 찾아야함.

+++ HoFramework/Define.h
@@ -35,6 +35,6 @@ struct Mesh
 {
 public:
 	vector<Vertex> vertices;
-	vector<uint32_t> indices;
+	vector<uint16_t> indices;
 	string textureSourceName;
 };
